# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CoffeeShop:
    properties:
      address:
        type: string
      city:
        type: string
      location:
        items:
          type: number
        type: array
      name:
        type: string
      rating:
        type: number
      roaster:
        type: boolean
    type: object
  models.EmptyBody:
    type: object
...
  models.FollowUnfollowRequest:
    properties:
      userFromId:
        type: string
      userToId:
        type: string
    type: object
  models.GetUserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isStaff:
        type: string
      lastName:
        type: string
      username:
        type: string
    type: object
...
  types.ApiError:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
host: go-coffee-api.vercel.app
info:
  contact:
    email: eduardozepeda@coffeebytes.dev
    name: Eduardo Zepeda
  description: This API returns information about speciality coffee shops in Guadalajara,
    Mexico.
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: http://swagger.io/terms/
  title: Coffee Shops in Gdl API
  version: "1.0"
paths:
  /coffee-shops:
    get:
      consumes:
      - application/json
      description: Get a list of all coffee shop in Guadalajara. Use page and size
        GET arguments to regulate the number of objects returned and the page, respectively.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Size number
        in: query
        name: size
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: User longitude
        in: query
        name: longitude
        type: number
      - description: User latitude
        in: query
        name: latitude
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CoffeeShop'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get a list of coffee shops
      tags:
      - coffee shops
    post:
      consumes:
      - application/json
      description: Create a coffee shop object.
      parameters:
      - description: New Coffee Shop data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CoffeeShop'
      - description: With the bearer started. Only staff members
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CoffeeShop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Create a new coffee shop
      tags:
      - coffee shops
  /coffee-shops/{coffee_shop_id}:
    delete:
      consumes:
      - application/json
      description: Delete a coffee shop object by its Id.
      parameters:
      - description: Coffee Shop ID
        in: path
        name: coffee_shop_id
        required: true
        type: string
      - description: With the bearer started. Only staff members
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.EmptyBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Delete a coffee shop
      tags:
      - coffee shops
    get:
      consumes:
      - application/json
      description: Get a specific coffee shop object. Id parameter must be an integer.
      parameters:
      - description: Coffee Shop ID
        in: path
        name: coffee_shop_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CoffeeShop'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get a coffee shop by its id
      tags:
      - coffee shops
    put:
      consumes:
      - application/json
      description: Update a coffee shop object by its Id.
      parameters:
      - description: Updated Coffee Shop data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CoffeeShop'
      - description: With the bearer started. Only staff members
        in: header
        name: Authorization
        required: true
        type: string
      - description: Coffee Shop ID
        in: path
        name: coffee_shop_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CoffeeShop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Update a coffee shop
      tags:
      - coffee shops
...
  /followers/{user_id}:
    get:
      consumes:
      - application/json
      description: Return user's followers from a given user Id
      parameters:
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetUserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Return user's followers,
      tags:
      - follows
  /following:
    post:
      consumes:
      - application/json
      description: Follow a user account using its id
      parameters:
      - description: Follow a user account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FollowUnfollowRequest'
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FollowUnfollowRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Follow user,
      tags:
      - follows
  /following/{user_id}:
    delete:
      consumes:
      - application/json
      description: Unfollow a user account using its id
      parameters:
      - description: Unfollow a user account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FollowUnfollowRequest'
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.FollowUnfollowRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Unfollow user,
      tags:
      - follows
    get:
      consumes:
      - application/json
      description: Return following users from a given user Id
      parameters:
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetUserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Return following users,
      tags:
      - follows
...
swagger: "2.0"